fix: 현재 작성자 기획에 맞게 수정 및 순환 참조 문제 해결 (#277)
